+++ circle-3/philosophers/philo/time.c
@@ -6,16 +6,16 @@
 /*   By: bel-mous <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/05 16:36:42 by bel-mous          #+#    #+#             */
-/*   Updated: 2022/05/12 14:21:38 by bel-mous         ###   ########.fr       */
+/*   Updated: 2022/05/26 19:07:15 by bel-mous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-int	get_timestamp_in_ms(long start_time)
+long	get_timestamp_in_ms(long start_time)
 {
 	struct timeval	time;
-	int				timestamp_in_ms;
+	long			timestamp_in_ms;
 
 	gettimeofday(&time, NULL);
 	timestamp_in_ms = (time.tv_sec * 1000) + (time.tv_usec) / 1000 - start_time;
